+++ mod.ts
@@ -4,7 +4,7 @@ export type {
 	EventHandlerFor,
 } from './src/arkiver/types.ts'
 export { Manifest } from './src/arkiver/manifest-builder.ts'
-export { ArkiveConsoleLogHandler, logger } from './src/logger.ts'
+export { ArkiveConsoleLogHandler } from './src/logger.ts'
 export { Store } from './src/arkiver/store.ts'
 export { Arkiver } from './src/arkiver/arkiver.ts'
 export { createEntity } from './src/graphql/entity.ts'